[Hotfix] 🍺 영업점 정보 거리 오류 해결

+++ src/main/java/com/hanaro/schedule_hanaro/customer/dto/response/BranchDetailResponse.java
@@ -36,7 +36,7 @@ public record BranchDetailResponse(
 	List<Integer> waitTime,
 
 	@JsonProperty("distance")
-	double distance
+	int distance
 ) {
 	public static BranchDetailResponse of(Long id, String branchName, String xPosition, String yPosition,
 		String address) {
